I fixed the find data method!
I added a view parameter, and this is what stopped the app from
crashing. Next time I am on, I plan to make it so that if there are
multiple instances of the same name, they all show up. I could also make
it so that all entries with the search word as part of the entry are
called. The extra credit option seems appealing too.